refactor(user-model): move user-model into infrastructure folder
